requirements.txt
anthropic>=0.40.0
aiohttp>=3.9.0
curl_cffi>=0.7.0
beautifulsoup4>=4.12.0
aiosqlite>=0.20.0
lxml>=5.0.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
dnspython>=2.6.0
pydantic>=2.0.0
